Create random_strong_digraph
Takes in the number of vertices and creates a strongly connected digraph. It utilizes the fact that each vertex in a  strongly connected digraph is contained in at least one directed cycle. It chooses the number of cycles (c) that should be present and it creates a random permutation of random length using the vertices "c" times. The adjacent vertices form the edges of the digraph.